requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.12.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
Bottleneck==1.6.0
build==1.4.0
CacheControl==0.14.4
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
cryptography==46.0.3
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
eval_type_backport==0.3.1
fastapi==0.128.0
filelock==3.20.3
firebase_admin==7.1.0
flatbuffers==25.12.19
fsspec==2026.1.0
google-api-core==2.28.1
google-auth==2.47.0
google-cloud-bigquery==3.40.0
google-cloud-core==2.5.0
google-cloud-firestore==2.22.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-genai==1.57.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gradio_client==2.0.3
groq==1.0.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
h5py==3.15.1
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.20.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
invoke==2.2.1
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
langchain-core==1.2.7
langdetect==1.0.9
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.2
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
numexpr==2.14.1
numpy==1.26.4
oauthlib==3.3.1
onnx==1.20.1
onnxruntime==1.23.2
opencv-python==4.8.1.78
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.1
overrides==7.7.0
packaging==24.2
pillow==10.4.0
pinecone==8.0.0
pinecone-plugin-assistant==3.0.1
pinecone-plugin-interface==0.0.7
posthog==5.4.0
proto-plus==1.27.0
protoc-gen-openapiv2==0.0.1
psutil==7.2.1
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pyclipper==1.4.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pypdfium2==4.30.0
PyPika==0.50.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-doctr==1.0.0
python-dotenv==1.2.1
python-multipart==0.0.6
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
scipy==1.17.0
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
sympy==1.13.1
tenacity==9.1.2
tokenizers==0.22.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.3.0
uuid_utils==0.13.0
uvicorn==0.40.0
validators==0.35.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
opencv-python==4.8.1.78
python-doctr[torch]
torch==2.5.1
torchvision==0.20.1
Pillow==10.4.0
numpy==1.26.4
groq>=0.11.0
reportlab>=4.0.0
svglib>=1.0.0
